==> 05-results.Rmd <==
# Results

```{r, include=FALSE}
#load packages
library(data.table)
library(dplyr)
library(lubridate)
library(scales)
library(ggplot2)
library(RColorBrewer)
library(tidyverse)
library(readr)
library(usdata)
library(ggridges)
library(Hmisc)
library(ggpubr)
library(stats)
library(patchwork)
library(GGally)
library(plotly)
library(gridExtra)
```

```{r, include=FALSE}
df <- read.csv("C:/Users/20961/Desktop/transformed_2008.csv")
```

## Flight Cancellation Analysis

**A=Carrier Caused, B=Weather, C=National Aviation System, D=Security**
```{r, include=FALSE}
#Q1 What is the most common reason for flight cancellation?
#A-Carrier Caused
#B-Weather
#C-National Aviation System
#D-Security
#https://aspm.faa.gov/aspmhelp/index/ASPM_Data_Download__Definitions_of_Variables.html
cancel_origin <- df %>%
  select(Origin_state, CancellationCode) %>%
  na.omit %>%
  group_by(Origin_state,CancellationCode) %>%
  mutate(n_cancel = n()) %>%
  unique()

cancel_origin2 <- cancel_origin %>%
  group_by(Origin_state) %>%
  mutate(total_cancel = sum(n_cancel)) %>%
  select(Origin_state, total_cancel) %>%
  unique() %>%
  arrange(desc(total_cancel))

cancel_origin2 <- cancel_origin2[1:10,]
```

```{r, include=FALSE}
unique(df$CancellationCode)
```

```{r,fig.height=6}
plot11 <- df %>% 
  filter(!is.na(CancellationCode)) %>%
  ggplot(aes(x = CancellationCode,na.rm = TRUE))+
  geom_bar(fill = "cornflowerblue")+
  xlab("Reason for Cancellation")+
  ylab("Count")+
  ggtitle("Main Reasons for Flight Cancellation")+
  theme(plot.title = element_text(size = 10))
  #scale_x_discrete(labels=c("Carrier Caused", "Weather","National Aviation System","Security"))

plot12 <- df %>% 
  filter(!is.na(CancellationCode)) %>%
  ggplot(aes(x = CancellationCode,na.rm = TRUE))+
  geom_bar()+
  xlab("Reason for Cancellation")+
  ylab("Count")+
  #scale_x_discrete(labels=c("Carrier Caused", "Weather","National Aviation System","Security"))+
  facet_wrap(~UniqueCarrier)+
  ggtitle("Number of Cancellation by Reason Faceted by Carrier")+
  theme(plot.title = element_text(size = 12))

plot13 <- cancel_origin2 %>%
  ggplot()+
  geom_bar(mapping = aes(x = reorder(Origin_state,-total_cancel), y = total_cancel), stat = "identity", fill = "lightblue")+
  xlab("Origin State")+
  ylab("Count")+
  ggtitle("Flight Cancellation over the U.S. States")+
  theme(plot.title = element_text(size = 10))

#ggarrange(ggarrange(plot11, plot13, ncol = 2), plot12, nrow = 2,heights = c(3,7)) 
```

```{r, fig.height=2}
grid.arrange(plot11, plot13, ncol=2)
```

Flight cancelation is one of the biggest concerns for travelers since it not only affects their original schedule but also increases the costs of the air carriers to accommodate the travelers. If a flight is canceled, the travelers have to take the next available same route flight, which may be the same airline on the next day due to the full occupancy of the same-day upcoming flights. In this way, the travelers’ schedules will be harmfully affected. On the other hand, in order to discharge the responsibilities and keep good customer experiences, the air carriers need to cover the rebooking costs, hotel fees, and other related fees. Therefore, finding the reasons for flight cancellation and looking at its pattern in the U.S. will help the travelers better understand the flight performances.

* Based on the bar plot on the left above, the number of flights caused by the air carrier is the highest. The air carrier's cause of the cancellation was due to circumstances within the airline's control (e.g. maintenance or crew problems, aircraft cleaning, baggage loading, fueling, etc.). Weather condition is another noticeable factor that causes flight cancellations. Extreme weather, significant meteorological conditions (actual or forecasted) that, in the judgment of the carrier, can prevent the operation of a flight such as a tornado, blizzard, or hurricane.

* Based on the bar plot on the right above, large airports or the airports in the main areas(e.g., Illinois, Texas, California, New York, etc.) of the U.S. are most likely to have the flights canceled.  According to both plots above, we may conclude that the air carriers that operate in the airports of those states have significant airline control problems or more extreme weather conditions than other states in the U.S. Since these states have some of the largest airports in the U.S., such as O'Hare International Airport, Dallas Fort Worth International Airport, and Los Angeles International Airport, it is reasonable to believe that the total number of flights is very large in those states.


```{r}
plot12
```

```{r, include=FALSE}
#test
test <- df[,c("Origin_state","CancellationCode")]
test <- test[!is.na(test$CancellationCode),]
test <- na.omit(test)
nrow(test[test$Origin_state == "CA",])
```

```{r, include=FALSE}
#filter and transform data
df2 <- df %>%
  select("UniqueCarrier","CancellationCode") %>%
  group_by(UniqueCarrier) %>%
  mutate(total_flight = n())

df22 <- df2 %>%
  na.omit() %>%
  select("UniqueCarrier","CancellationCode") %>%
  group_by(UniqueCarrier,CancellationCode) %>%
  mutate(n_cancel = n()) %>%
  unique() %>%
  arrange(UniqueCarrier,CancellationCode)

df23 <- df22 %>%
  group_by(UniqueCarrier) %>%
  mutate(total_cancel = sum(n_cancel))

df2 <- df2 %>%
  select("UniqueCarrier","total_flight") %>%
  unique()

df24 <- merge(df23,df2,by="UniqueCarrier",all.x = TRUE)
df24 <- df24 %>%
  mutate(cancel_rate = total_cancel/total_flight)

getPalette = colorRampPalette(brewer.pal(4, "Purples"))
```

```{r, fig.height=4}
ggplot(df24, aes(fill=fct_rev(CancellationCode), y=n_cancel, x=reorder(UniqueCarrier,-total_cancel))) +
  geom_bar(position="stack", stat="identity") +
  scale_fill_manual(values=getPalette(5))+
  geom_line(aes(x=reorder(UniqueCarrier,-total_cancel), y = 100000*cancel_rate), size = 1, color="red", group = 1) + 
  scale_y_continuous(sec.axis = sec_axis(~./100000, name = "Cancellation Rate"))+
  ggtitle("Air Carrier Performance on Flight Cancellation")+
  xlab("Carrier")+
  ylab("Number of Cancellation")+
  labs(fill = "Cancellation Code")
```

Travelers may also consider which air carrier to choose to minimize the probability of flight cancellation when those carriers provide similar services and prices. From the plots above, American Airlines, Envoy Air, and SkyWest Airlines have the most flight cancellations. American Airlines is the airline with the largest passenger capacity, total passenger mileage, and fleet in the world. Therefore, we may conclude that the total number of flights is very large so the total number of flight cancellations can be large too and it is hard to manage a large number of flights. However, extreme weather and National Aviation System (NAS) are big problems for Envoy Air and SkyWest Airlines, where NAS cancellations are attributable to the national aviation system that refer to a broad set of conditions, such as non-extreme weather conditions, airport operations, heavy traffic volume, and air traffic control.

## Flight Delay Analysis
```{r, include=FALSE}
#delay
delay <- df %>%
  select(c("DepDelay","CarrierDelay","WeatherDelay","NASDelay","SecurityDelay","LateAircraftDelay"))

delay <- delay[delay$DepDelay > 0,]

delay2 <- delay %>% na.omit()
dc <- nrow(delay2[delay2$CarrierDelay > 0,])
dw <- nrow(delay2[delay2$WeatherDelay > 0,])
dn <- nrow(delay2[delay2$NASDelay > 0,])
ds <- nrow(delay2[delay2$SecurityDelay > 0,])
dl <- nrow(delay2[delay2$LateAircraftDelay > 0,])

c1 <- c("CarrierDelay","WeatherDelay","NASDelay","SecurityDelay","LateAircraftDelay")

c2 <- c(dc,dw,dn,ds,dl)

delay_count <- data.frame("reason" = c1, "count" = c2, stringsAsFactors = FALSE)
```

```{r, include=FALSE}
df31 <- df[,c1] %>%
  na.omit() 
df32 <- df31 %>%
  mutate(id = row_number()) %>%
  pivot_longer(!id, names_to = "reason", values_to = "delay_min") %>%
  select("reason","delay_min") %>%
  arrange(reason)

df32 <- df32[df32$delay_min != 0,]

df33 <- df32 %>%
  group_by(reason) %>%
  mutate(avg_delay = mean(delay_min)) %>%
  select("reason","avg_delay") %>%
  unique()

getPalette2 = colorRampPalette(brewer.pal(5, "GnBu"))
```

```{r, include=FALSE}
#barchart: number of delay
plot31 <- delay_count %>%
  ggplot()+
  geom_bar(mapping = aes(x = reorder(reason,-count), y = count,fill = reason), stat = "identity")+
  ggtitle("Main Reasons for Flight Delay")+
  xlab("Reason for Delay")+
  ylab("Count")+
  scale_fill_manual(values=getPalette2(5))+
  theme(legend.position = "None",
        axis.text.x = element_text(size=6))

#barchart: average delay (mins)
plot32 <- df33 %>% ggplot()+
  geom_bar(mapping = aes(x = reorder(reason,-avg_delay), y = avg_delay, fill = reason), stat = "identity")+
  xlab("Reason for Delay")+
  ylab("Average Delay (mins)")+
  scale_fill_manual(values=getPalette2(5))+
  theme(legend.title = element_blank(),legend.position = "top",
        axis.text.x = element_text(size=6))
plot32_<- plot32+theme(legend.position = "None")+ggtitle("Main Reasons for Flight Delay")

#boxplot: delay reason vs delay mins
#explore median, outliers
plot33 <- ggplot(df32, aes(x=reason, y=delay_min)) + 
    geom_boxplot(notch=TRUE,notchwidth = 0.8,
        outlier.colour="red",outlier.size = 2,outlier.shape = 1)+
  xlab("Reason for Delay")+
  ylab("Delay (mins)")

#redgeplot: delay reason vs. delay mins
#explore distribution
plot34 <- df32[df32$delay_min <= 200,]%>%
  ggplot(aes(delay_min, fct_reorder(reason, delay_min,median))) + 
  geom_density_ridges(fill = "cornflowerblue", alpha = .3,scale = 1)+
  labs(x = "Delay (mins)",
       y = "Reason for Delay",
       caption = "Select delay less than 200 mintues")

#ggpubr::ggarrange(
#  ggarrange(legend, 
#            ggarrange(plot31, plot32_, ncol = 2, nrow = 1),
#            ncol = 1,nrow = 2,heights = c(2,10)),
#  plot33,plot34,ncol = 1, nrow = 3)
```


```{r, fig.height=2}
shared_legend <- get_legend(plot32)
grid.arrange(shared_legend,
             arrangeGrob(plot31, plot32_, ncol = 2),nrow = 2, heights = c(1, 10))
```

Even though flight delay may not be as influential as flight cancellation, it still bothers millions of travelers everyday. Here, we would like to analyze the main reason of the flight delays and the average delay time for those delays in 2008.

* Based on the bar plot on the left above, delay attributable to the National Aviation System, delay attributable to the late-arriving aircraft, and delay attributable to the carrier caused most of the flight delays.

* Based on the bar plot on the right above, the average delay time of delay attributable to the weather,  delay attributable to the late-arriving aircraft, and delay attributable to the carrier is the highest.


```{r,fig.height=2}
plot33+ggtitle("Distributions of Flight Delay Reasons")
```

```{r,fig.height=2}
plot34
```

Also, as shown in the boxplot and ridgeline plot above, the average delay time distributes most from 10 minutes to 25 minutes.

Therefore, we highly recommend the travelers check the weather condition before traveling, choose reliable airlines that have good management systems, and avoid the flights that always have late previous airlines.

```{r, include=FALSE}
#Top 10 carrier which occur delays most

#airline/carrier
#Top 10 airlines measured by average delay
carrier1 <- df %>%
  select(c('UniqueCarrier','DepDelay'))%>%
  group_by(UniqueCarrier) %>%
  mutate(avg_delay = round(mean(DepDelay, na.rm = TRUE),digits = 2)) %>%
  select(c('UniqueCarrier','avg_delay')) %>%
  unique() %>%
  arrange(desc(avg_delay))
```

```{r, include=FALSE}
#airline/carrier
#Top 10 airlines measured by percentage of delayed flights
car_del <- df %>%
  select(c('UniqueCarrier', 'DepDelay')) %>%
  mutate(DelayOrNot = case_when(
    DepDelay > 0 ~ 1,
    DepDelay <= 0 ~ 0
  ))

pct_delayed <- car_del %>%
  group_by(UniqueCarrier) %>%
  na.omit() %>%
  mutate(total_flight = n(), total_delayed = sum(DelayOrNot)) %>%
  select(c('UniqueCarrier','total_flight','total_delayed')) %>%
  unique() %>%
  mutate(pct = total_delayed/total_flight) %>%
  arrange(desc(pct))

df41 <- merge(carrier1,pct_delayed, by="UniqueCarrier",all.x = TRUE)
df41 <- df41 %>%
  mutate(judge = case_when(
    avg_delay > 0 ~ TRUE,
    avg_delay < 0 ~ FALSE
  ))
```

```{r, include=FALSE}
#tmp3 -> df42
df42 <- df %>%
  select(c('UniqueCarrier','CarrierDelay','WeatherDelay','NASDelay','SecurityDelay','LateAircraftDelay')) %>%
  na.omit() %>%
  mutate(carrier = case_when(
    CarrierDelay > 0 ~ 1,
    CarrierDelay <= 0 ~ 0
  ), weather = case_when(
    WeatherDelay > 0 ~ 1,
    WeatherDelay <= 0 ~ 0
  ), nas = case_when(
    NASDelay > 0 ~ 1,
    NASDelay <= 0 ~ 0
  ), security = case_when(
    SecurityDelay > 0 ~ 1,
    SecurityDelay <= 0 ~ 0
  ), late = case_when(
    LateAircraftDelay > 0 ~ 1,
    LateAircraftDelay <= 0 ~ 0
  ))

df42 <- df42 %>%
  select(c('UniqueCarrier','carrier','weather','nas','security','late'))

df42 <- df42 %>%
  group_by(UniqueCarrier) %>%
  mutate(total_carrier = sum(carrier),
         total_weather = sum(weather),
         total_nas = sum(nas),
         total_security = sum(security),
         total_late = sum(late)) %>%
  select(c('UniqueCarrier','total_carrier','total_weather','total_nas','total_security','total_late')) %>%
  unique()

df42$total = rowSums(df42[-1])
df42_p <- df42[,c("UniqueCarrier","total")]

#tmp4 -> df43
df42 <- df42[-c(7)]
df43 <- df42 %>%
  pivot_longer(!UniqueCarrier, names_to = "reason", values_to = "num_delay")

#tmp5 -> df44
df44 <- merge(df43,df42_p,y.by = "UniqueCarrier")

getPalette3 = colorRampPalette(brewer.pal(6, "Set2"))
```

```{r, include=FALSE}
plot41 <- ggplot(df41, aes(x = reorder(UniqueCarrier,-avg_delay), y = avg_delay, fill = judge))+
  geom_bar(stat = "identity")+
  geom_line(aes(x = reorder(UniqueCarrier,-avg_delay), y = pct*15), color = "red", size = 0.9, group = 1, alpha = 0.6)+
  geom_point(aes(x = reorder(UniqueCarrier,-avg_delay), y = pct*15), size = 1.5, group = 1, color = "red", fill = "red", alpha = 0.6)+
  scale_y_continuous(sec.axis=sec_axis(~./15,name = "Percentage of Delayed Flights"))+
  scale_fill_manual(values = c("#92c0df","#fcbe75"))+
  theme(legend.position = "None")+
  ggtitle("Air Carrier Performance on Flight Delay")+
  ylab("Average Delay (mins)")+
  xlab("Carrier")

plot42 <- ggplot(df44, aes(fill=fct_rev(reason), y=num_delay, x=reorder(UniqueCarrier,total))) + 
    geom_bar(position="stack", stat="identity")+
  xlab("Airline")+
  ylab("Total Number of Delay")+
  scale_fill_manual(values=getPalette3(5),name = "Reason", labels = c("Weather", "Security", "NAS","Late Aircraft","Carrier"))+
  coord_flip()

#ggarrange(plot41,plot42,ncol = 1, nrow = 2, heights = c(3,5))
```

```{r, fig.height=7}
grid.arrange(plot41,plot42,nrow = 2, heights = c(3, 5))
```

Travelers may also consider which air carrier to choose to minimize the probability of flight delay when those carriers provide similar services and prices.

* Based on the first plot above, United Airlines, American Airlines, Eva Air, Envoy Air, Mesa Airlines, and Continental Airlines have the longest average delay time that is above 12.5 minutes. Thus, according to the delay reasons discussed before, we suggest those air carriers improve their airline controls and flight on-time management system. It is noticeable that Hawaiian Airlines and 9 Air have negative average delay time, which implies that their flights are very likely to arrive early than the scheduled time. Thus, these two airlines are good choices for travelers.

* Based on the second plot above, Southwest Airlines, American Airlines, Envoy Air, and United 	Airlines have the largest total number of delay, in which Southwest Airlines has extremely a large amount of flights delayed. NAS and Late Aircraft Arrival are the biggest problems that caused those delays, so airport operations, heavy traffic volume, and air traffic control may be the root causes. It can be easily observed that United Airlines and American Airlines have very long average delay time and a large number of delays.

Besides the price, time, and availability, travelers usually choose the air carriers that have the best service, minimum delay or cancellation probability, and best reputation. According to the [Airline Ranking 2008 published by the Bureau of Transportation Statistics](https://www.bts.gov/airline-rankings-2008),  United Airlines and American Airlines are the largest U.S. air carriers in terms of revenue, the number of passengers, and miles. Even though the performances of United Airlines and American Airlines are not that good based on our observations,  we assume the total number of flights of those two carriers is very large so the total number of flight delays can be large too and it is hard to manage a large number of flights. 

```{r, include=FALSE}
usa <- df %>%
  select("DepDelay","Origin_state") %>%
  na.omit() %>%
  group_by(Origin_state) %>%
  mutate(avg_delay = mean(DepDelay)) %>%
  select("Origin_state","avg_delay") %>%
  unique()
```

```{r}
l <- list(color = toRGB("white"), width = 2)
# specify some map projection/options
g <- list(
  scope = 'usa',
  projection = list(type = 'albers usa'),
  showlakes = TRUE,
  lakecolor = toRGB('white')
)

fig <- plot_geo(usa, locationmode = 'USA-states')
fig <- fig %>% add_trace(
  z = usa$avg_delay, locations = usa$Origin_state,
  color = usa$avg_delay, colors = 'GnBu'
)
fig <- fig %>% colorbar(title = "Average Delay (mins)")
fig <- fig %>% layout(
  title = '2008 US Flights Depature Delay by State<br>(Measured by Average Delay in Minutes)',
  geo = g
)

fig
```

As we can see from the spatial plot above, Illinois and New Jersey rank highest nationwide for departure delays. The main airport in Illinois is O'Hare International Airport, the busiest airport in the U.S., and Newark Liberty Airport is also ranked in the U.S.’s 29 busiest airports for departures. Based on our previous observations that Air Carriers, NAS, and Late Aircraft Arrival are the biggest problems for flight delays. Therefore, the following insights can be obtained:

* O'Hare International Airport and Newark Liberty Airport are unquestionably the busiest airspace, so any one delay really sends those airports into a disadvantage for on-time performance

* Due to the increase in population and airlines,  the airports are operating at capacities larger than they were built for

```{r, include=FALSE}
df61 <- df[,c("DepDelay","Origin")] %>%
  na.omit() %>%
  group_by(Origin) %>%
  mutate(avg_delay = mean(DepDelay)) %>%
  select("Origin","avg_delay") %>%
  unique() %>%
  arrange(desc(avg_delay))

df61_2 <- df61 %>%
  arrange(avg_delay)

org_head10 <- df61[1:10,]
org_tail10 <- df61_2[1:10,]

org <- rbind(org_head10, org_tail10)
org["judge"] = org$avg_delay > 0

df62 <- df[,c("ArrDelay","Dest")] %>%
  na.omit() %>%
  group_by(Dest) %>%
  mutate(avg_delay = mean(ArrDelay)) %>%
  select("Dest","avg_delay") %>%
  unique() %>%
  arrange(desc(avg_delay))

df62_2 <- df62 %>%
  arrange(avg_delay)

des_head10 <- df62[1:10,]
des_tail10 <- df62_2[1:10,]

des <- rbind(des_head10, des_tail10)
des["judge"] = des$avg_delay > 0
```

```{r, include=FALSE}
plot61 <- ggplot(org, aes(x = reorder(Origin, -avg_delay), y = avg_delay, fill = judge))+
  geom_bar(stat = "identity")+
  scale_fill_manual(values = c("#92c0df","#fcbe75"))+
  theme(legend.position = "None")+
  ggtitle("Airport Performance on Flight Cancellation")+
  ylab("Average Departure Delay (mins)")+
  xlab("Origin Airport")

plot62 <- ggplot(des, aes(x = reorder(Dest, -avg_delay), y = avg_delay, fill = judge))+
  geom_bar(stat = "identity")+
  scale_fill_manual(values = c("#92c0df","#fcbe75"))+
  theme(legend.position = "None")+
  ylab("Average Arrival Delay (mins)")+
  xlab("Destination Airport")

#ggarrange(plot61,plot62,ncol = 1,nrow = 2)
```

```{r, fig.height=5}
grid.arrange(plot61,plot62,nrow = 2)
```

The two plots above show the top average departure delay of the origin airport and the average arrival delay of the destination airport. An interesting observation is that some airports, including Hilo International Airport,  Lihue Airport, Bert Mooney Airport, Bellingham International Airport, and Tupelo Regional AirportTupelo Regional Airport, have negative average departure delay time and negative average arrival delay time. This implies that those airports have better airline management systems and have airlines always ahead of schedule.

```{r, include=FALSE}
df71 <- df %>%
  filter(UniqueCarrier == "WN"|UniqueCarrier == "CO"|UniqueCarrier == "F9"|UniqueCarrier == "AA"|UniqueCarrier == "UA"|UniqueCarrier == "EV"|UniqueCarrier == "MQ"|UniqueCarrier == "XE"|UniqueCarrier == "FL"|UniqueCarrier == "AS") %>%
  select(c('UniqueCarrier','DepHour','DepDelay'))

df71 <- df71 %>%
  mutate(DelayOrNot = case_when(
    DepDelay > 0 ~ 1,
    DepDelay <= 0 ~ 0
  ))

df71 <- df71 %>%
  group_by(UniqueCarrier, DepHour) %>%
  mutate(total_flight_hour = n(), total_delay_hour = sum(DelayOrNot)) %>%
  mutate(pct_delay = total_delay_hour/total_flight_hour)

df72 <- df71 %>%
  select(c(UniqueCarrier,DepHour,pct_delay)) %>%
  unique() %>%
  arrange(UniqueCarrier,DepHour)%>%
  na.omit()
```

```{r}
df73 <- df %>%
  filter(UniqueCarrier == "WN"|UniqueCarrier == "CO"|UniqueCarrier == "F9"|UniqueCarrier == "AA"|UniqueCarrier == "UA") %>%
  select(c('UniqueCarrier',"DayOfWeek",'DepHour','DepDelay'))

df73 <- df73 %>%
  mutate(DelayOrNot = case_when(
    DepDelay > 0 ~ 1,
    DepDelay <= 0 ~ 0
  ))

df73 <- df73 %>%
  group_by(UniqueCarrier,DayOfWeek,DepHour) %>%
  mutate(total_flight_hour = n(), total_delay_hour = sum(DelayOrNot)) %>%
  mutate(pct_delay = total_delay_hour/total_flight_hour)

df74 <- df73 %>%
  select(c(UniqueCarrier,DayOfWeek,DepHour,pct_delay)) %>%
  unique() %>%
  arrange(UniqueCarrier,DayOfWeek,DepHour)%>%
  na.omit() %>%
  mutate(week = case_when(
    DayOfWeek == 1 ~ "Monday",
    DayOfWeek == 2 ~ "Tuesday",
    DayOfWeek == 3 ~ "Wednesday",
    DayOfWeek == 4 ~ "Thursday",
    DayOfWeek == 5 ~ "Friday",
    DayOfWeek == 6 ~ "Saturday",
    DayOfWeek == 7 ~ "Sunday"
  ))

df74$week <- factor(df74$week,levels = c("Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"))
```

```{r}
ggplot(data = df72 ) +
  geom_tile(aes(x = DepHour, y = UniqueCarrier, fill = pct_delay), color = 'black') +
  scale_x_continuous(breaks = c(5,11,17,22), labels = function(x){
    case_when(x == 5 ~ '5am', x == 11 ~ '11am', x == 17 ~ '5pm', x == 22 ~ '10pm')
  }) + coord_equal()+
  scale_fill_distiller(palette = 'Spectral') +
  labs(x = 'Hour of day', y = 'Airline', fill = 'Percentage',
       title = 'Distribution of Flight Delay Over A Day') +
  theme(panel.grid.major = element_blank(), axis.ticks = element_blank(),
        plot.caption = element_text(vjust = 7), axis.title.y = element_blank())

```

If we have choices, what is the best time of the day to start our trip? The heatmap above answers this question. It is obvious that the earlier in the day we fly, the less likely we’ll be stuck with flight delays, and delay times only get worse as the day goes on. The probable reasons are the following：

* The weather conditions at the night are worse. The storms, hurricanes, and heavy snow are more likely to happen at the night

* There are too many flight delays during the daytime, and airline traffic and airport operation problems are getting worse at the end of the day.

```{r,include=FALSE}
#test
test <- df[,c("UniqueCarrier","DepHour","DepDelay")]
test <- na.omit(test)
sum(test[test$UniqueCarrier == "AA" & test$DepHour == 2,]$DepDelay > 0)/nrow(test[test$UniqueCarrier == "AA" & test$DepHour == 2,])
```

```{r, fig.height=3}
ggplot(data = df74) +
  geom_line(aes(x = DepHour, y = pct_delay, color = UniqueCarrier), size = 0.5) +
  geom_point(aes(x = DepHour, y = pct_delay, color = UniqueCarrier), size = 0.7) +
  facet_grid(~week) +
  scale_x_continuous(limits = c(0,24), breaks = 8*c(0:3), labels = function(x){
    case_when(x == 0 ~ '', x == 8 ~ '8am', x == 16 ~ '4pm', x == 24 ~ '12am')
    }) +
  labs(x = 'Time', y = 'Percentage of delayed flights', color = 'Airline',
       title = 'Distribution of Flight Delay Over A Week',
       subtitle = 'Top US airlines') +
  theme(plot.caption = element_text(vjust = 7)) +
  guides(fill = guide_legend(title = NULL))
```

```{r, include=FALSE}
df76 <- df[,c("UniqueCarrier","DepDelay","DayOfWeek")] %>%
  mutate(week = case_when(
    DayOfWeek == 1 ~ "Mon",
    DayOfWeek == 2 ~ "Tue",
    DayOfWeek == 3 ~ "Wed",
    DayOfWeek == 4 ~ "Thur",
    DayOfWeek == 5 ~ "Fri",
    DayOfWeek == 6 ~ "Sat",
    DayOfWeek == 7 ~ "Sun"
  )) %>%
  na.omit()

df76$week <- factor(df76$week,levels = c("Sun", "Mon", "Tue", "Wed", "Thur", "Fri", "Sat"))
```

```{r, fig.height=8}
df76[,c("UniqueCarrier","week","DepDelay")] %>% 
  na.omit() %>%
  ggplot(aes(x = week, y = DepDelay, fill = week))+
  geom_violin() + 
  facet_wrap(~UniqueCarrier,ncol = 4) +
  xlab("Day of Week")+
  ylab("Departure Delay (mins)")
```

According to the plots above, the delays spread over 24 hours of a day and 7 days of a week, so there is no noticeable relationship between ‘Time of a day’ and ‘Total number of delays’. Also, there is no noticeable relationship between ‘Day of a week’ and ‘Total departure delay time’. Therefore, there is no seasonality to the delays by airlines.

## Exploratory Analysis
```{r, include=FALSE}
df91 <- df %>%
  select(c("Distance", "DepDelay","ArrDelay","AirTime","DepHour","DayofMonth","DayOfWeek")) %>%
  na.omit()
split = sample(x = 1:nrow(df91), size = 0.001*nrow(df91))
small_dist = df91[split,]
```

```{r, fig.height=6.5}
lowerFn <- function(data, mapping, ...) {
  p <- ggplot(data = data, mapping = mapping) +
    geom_point(color = 'blue', alpha=0.3, size=0.5) +
    geom_smooth(color = 'black', method='lm', size=0.5,...)
  p
}

g <- ggpairs( 
  data = small_dist,
  lower = list(
    continuous =  wrap(lowerFn) #wrap("smooth", alpha = 0.3, color = "blue", lwd=1) 
  ),
  upper = list(continuous = wrap("cor", size = 3))
)
g <- g + theme(
  axis.text = element_text(size = 8),
  axis.title = element_text(size = 8),
  strip.text.x = element_text(size = 8),
  strip.text.y = element_text(size = 8),
  legend.background = element_rect(fill = "white"),
  panel.grid.major = element_line(colour = NA),
  panel.grid.minor = element_blank(),
  panel.background = element_rect(fill = "grey95")
)+
  ggtitle("Correlation Analysis of Main Variables")

print(g, bottomHeightProportion = 0.5, leftWidthProportion = .5)
```

Last but not least, it is worthy to analyze the correlations between the main variables in our dataset. Are there any important insights we can obtain from those variables to help travelers better plan for their future trips? The scatterplot matrix above answers the question and shows:

* There is a positive correlation between

  * Departure Delay and Arrival Delay

  * Distance of the flight and the Arrival time

We assume that If the coefficient value lies between ± 0.50 and ± 1, then it is said to be a strong correlation. If the value lies between ± 0.30 and ± 0.49, then it is said to be a medium correlation. If the value lies below + 0.29, then it is said to be a small correlation. Thus, Arrival Time and Distance has a strong correlation, Departure Delay and Arrival Delay has a strong correlation, Departure Hour and Departure Delay has a low correlation, Departure Hour and Arrival Delay has a low correlation.

* No variable has a normal distribution
