==> 07-conclusion.Rmd <==
# Conclusion

After the flight cancelation analysis and flight delay analysis, we have gained a more thorough understanding of the U.S. airline on-time performance and obtained many insightful findings to answer the questions we proposed in the beginning. For example, we found the most common reasons for flight cancellations and delays, the relatively busier airports in 2008, and the possible airline control and airport management problems.

Despite these findings, our analysis is limited. The goal of our analysis was to help nowadays travelers have better insights into the situation they faced and the challenges they may meet in the future. Also, we desired to provide constructive suggestions for their plan. However, big data can tell a reliable story but still have some limitations. 

First, we used part of the U.S. airline on-time performance data of 2008, which shows important observations but additional data may be needed for more insightful analysis. Under the background of the COVID-19 pandemic since 2019, the airline industry was strictly harmed and adjusted to a large extent. In order to provide more up-to-date information to the travelers, we think it is necessary to gain the newest data from the Bureau of Transportation Statistics and perform the data analysis. 

In addition to limitations in data, there also existed limitations in analysis. There are lots of factors determining the overall performance of the airlines, not only flight cancellation rates and flight delays. Airline Quality Rating releases report each year to assess airline quality on combined multiple performance criteria. Also, customer experiences and ratings to the airlines are also important. In the future, a more comprehensive analysis can be helpful for travelers, transportation bureaus, airport managers, etc. 

In 2008, we found the busy scenario of the airports, growing airline industries, and the uncontrollable situations that influence the airline industry. At the beginning of the COVID-19 pandemic, we saw rearrangements and massive airline cancellations. Now, the U.S. is on its way to leaving the pandemic behind. For instance, borders reopen, people start to travel again, and the number of flights is constantly growing. Therefore, airlines’ on-time performance is back in focus. We hope to let the world know the importance of airline analysis and we are looking forward to having more precise insights in this big data era.
